+++ packages/outline-view/src/browser/outline-view-contribution.ts
@@ -16,13 +16,24 @@
 
 import { injectable } from 'inversify';
 import { AbstractViewContribution } from '@theia/core/lib/browser/shell/view-contribution';
-import { OutlineViewWidget } from './outline-view-widget';
 import { FrontendApplicationContribution, FrontendApplication } from '@theia/core/lib/browser/frontend-application';
+import { Command, CommandRegistry } from '@theia/core/lib/common/command';
+import { TabBarToolbarContribution, TabBarToolbarRegistry } from '@theia/core/lib/browser/shell/tab-bar-toolbar';
+import { Widget } from '@theia/core/lib/browser/widgets';
+import { OutlineViewWidget } from './outline-view-widget';
+import { CompositeTreeNode } from '@theia/core/lib/browser/tree';
 
 export const OUTLINE_WIDGET_FACTORY_ID = 'outline-view';
 
+export namespace OutlineViewCommands {
+    export const COLLAPSE_ALL: Command = {
+        id: 'outlineView.collapse.all',
+        iconClass: 'collapse-all'
+    };
+}
+
 @injectable()
-export class OutlineViewContribution extends AbstractViewContribution<OutlineViewWidget> implements FrontendApplicationContribution {
+export class OutlineViewContribution extends AbstractViewContribution<OutlineViewWidget> implements FrontendApplicationContribution, TabBarToolbarContribution {
 
     constructor() {
         super({
@@ -39,5 +50,41 @@ export class OutlineViewContribution extends AbstractViewContribution<OutlineVie
 
     async initializeLayout(app: FrontendApplication): Promise<void> {
         await this.openView();
+    }
+
+    registerCommands(commands: CommandRegistry): void {
+        super.registerCommands(commands);
+        commands.registerCommand(OutlineViewCommands.COLLAPSE_ALL, {
+            isEnabled: widget => this.withWidget(widget, () => true),
+            isVisible: widget => this.withWidget(widget, () => true),
+            execute: () => this.collapseAllItems()
+        });
+    }
+
+    registerToolbarItems(toolbar: TabBarToolbarRegistry): void {
+        toolbar.registerItem({
+            id: OutlineViewCommands.COLLAPSE_ALL.id,
+            command: OutlineViewCommands.COLLAPSE_ALL.id,
+            tooltip: 'Collapse All',
+            priority: 0
+        });
+    }
+
+    /**
+     * Collapse all nodes in the outline view tree.
+     */
+    protected async collapseAllItems(): Promise<void> {
+        const { model } = await this.widget;
+        const root = model.root;
+        if (CompositeTreeNode.is(root)) {
+            model.collapseAll(root);
+        }
+    }
+
+    protected withWidget<T>(widget: Widget | undefined = this.tryGetWidget(), cb: (widget: OutlineViewWidget) => T): T | false {
+        if (widget instanceof OutlineViewWidget && widget.id === OUTLINE_WIDGET_FACTORY_ID) {
+            return cb(widget);
+        }
+        return false;
     }
 }