+++ packages/outline-view/src/browser/outline-view-frontend-module.ts
@@ -25,12 +25,16 @@ import {
     bindViewContribution,
     TreeProps,
     defaultTreeProps,
-    TreeDecoratorService
+    TreeDecoratorService,
+    TreeModel,
+    TreeModelImpl
 } from '@theia/core/lib/browser';
+import { TabBarToolbarContribution } from '@theia/core/lib/browser/shell/tab-bar-toolbar';
 import { OutlineViewWidgetFactory, OutlineViewWidget } from './outline-view-widget';
 import '../../src/browser/styles/index.css';
 import { bindContributionProvider } from '@theia/core/lib/common/contribution-provider';
 import { OutlineDecoratorService, OutlineTreeDecorator } from './outline-decorator-service';
+import { OutlineViewTreeModel } from './outline-view-tree';
 
 export default new ContainerModule(bind => {
     bind(OutlineViewWidgetFactory).toFactory(ctx =>
@@ -42,6 +46,7 @@ export default new ContainerModule(bind => {
 
     bindViewContribution(bind, OutlineViewContribution);
     bind(FrontendApplicationContribution).toService(OutlineViewContribution);
+    bind(TabBarToolbarContribution).toService(OutlineViewContribution);
 });
 
 function createOutlineViewWidget(parent: interfaces.Container): OutlineViewWidget {
@@ -51,6 +56,10 @@ function createOutlineViewWidget(parent: interfaces.Container): OutlineViewWidge
 
     child.unbind(TreeWidget);
     child.bind(OutlineViewWidget).toSelf();
+
+    child.unbind(TreeModelImpl);
+    child.bind(OutlineViewTreeModel).toSelf();
+    child.rebind(TreeModel).toService(OutlineViewTreeModel);
 
     child.bind(OutlineDecoratorService).toSelf().inSingletonScope();
     child.rebind(TreeDecoratorService).toDynamicValue(ctx => ctx.container.get(OutlineDecoratorService)).inSingletonScope();